requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anykeystore=0.2=pypi_0
apex=0.9.10.dev0=pypi_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39hb9d737c_1004
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7f98852_4
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2023.5.7=py39h06a4308_0
cffi=1.14.6=py39he32792d_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=3.1.0=pyhd8ed1ab_0
click=8.0.4=py39h06a4308_0
cloudpickle=2.2.1=py39h06a4308_0
contourpy=1.0.5=py39hdb19cb5_0
cryptacular=1.6.2=pypi_0
cryptography=37.0.2=py39hd97740a_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py39h5eee18b_0
dask-core=2023.4.1=py39h06a4308_0
dbus=1.13.18=hb2f20db_0
defusedxml=0.7.1=pypi_0
expat=2.4.9=h6a678d5_0
fast-slic=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h0708190_1
fsspec=2023.4.0=py39h06a4308_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
greenlet=2.0.2=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
hupper=1.12=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
imagecodecs=2021.8.26=py39hf0132c2_1
imageio=2.26.0=py39h06a4308_0
importlib-metadata=6.0.0=py39h06a4308_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h166bdaf_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7f98852_1001
lazy_loader=0.1=py39h06a4308_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libcurl=7.88.1=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libllvm10=10.0.1=hbcb73fb_5
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=h21135ba_2
libpq=12.9=h16c4e8d_3
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=hecacb30_2
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py39h06a4308_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=py39h06a4308_1
matplotlib-base=3.7.1=py39h417a72b_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
networkx=2.8.4=py39h06a4308_1
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.24.3=py39h14f4228_0
numpy-base=1.24.3=py39h31eccc5_0
oauthlib=3.2.2=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.7.0.72=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1t=h7f8727e_0
packaging=23.0=py39h06a4308_0
partd=1.2.0=pyhd3eb1b0_1
pastedeploy=3.0.1=pypi_0
pbkdf2=1.3=pypi_0
pcre=8.45=h295c915_0
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
plaster=1.1.2=pypi_0
plaster-pastedeploy=1.0.1=pypi_0
ply=3.11=py39h06a4308_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pyramid=2.0.1=pypi_0
pyramid-mailer=0.15.1=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.0=hdb3f193_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python3-openid=3.2.0=pypi_0
python_abi=3.9=2_cp39
pytorch=1.12.0=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=py39h5eee18b_0
pyyaml=6.0=py39h5eee18b_1
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
repoze-sendmail=4.4.1=pypi_0
requests=2.29.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
scikit-image=0.20.0=py39h6a678d5_0
scikit-learn=1.2.2=py39h6a678d5_1
scipy=1.9.1=py39h14f4228_0
setuptools=66.0.0=py39h06a4308_0
sip=6.6.2=py39h6a678d5_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h295c915_0
sqlalchemy=1.4.48=pypi_0
sqlite=3.41.2=h5eee18b_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
timm=0.3.2=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py39h06a4308_0
torchaudio=0.12.0=py39_cu116
torchprofile=0.0.4=pypi_0
torchvision=0.13.0=py39_cu116
tornado=6.2=py39h5eee18b_0
tqdm=4.65.0=py39hb070fc8_0
transaction=3.1.0=pypi_0
translationstring=1.4=pypi_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=pyhd8ed1ab_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
webob=1.8.7=pypi_0
wheel=0.38.4=py39h06a4308_0
wtforms=3.0.1=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zfp=0.5.5=h295c915_6
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zope-deprecation=5.0=pypi_0
zope-interface=6.0=pypi_0
zope-sqlalchemy=2.0=pypi_0
zstd=1.5.2=ha4553b6_0
